Move customConfigItemGroup components to new folder #2489
